+++ SerialPrograms/Source/PokemonLA/Inference/Map/PokemonLA_MapMarkerLocator.h
@@ -28,7 +28,7 @@ enum class MapRegion;
 //   It also does not work if the player is near the volcano area when viewing the map in zoom level
 //   2 (local area view). If `avoid_lava_area` is true, skip the red pixels that are in the upper right
 //   area of the map, which is well the lava area is on the coastlands map at zoom level 1.
-float get_orientation_on_map(const ImageViewRGB32& screen, bool avoid_lava_area = false);
+double get_orientation_on_map(const ImageViewRGB32& screen, bool avoid_lava_area = false);
 
 
 }
